substack-alerter/requirements.txt
discord.py==1.7.3
feedparser==6.0.8
python-dotenv==0.19.2
python_dateutil==2.8.2
SQLAlchemy==1.4.31
